- The account edit dialog now shows only one dialog containing all the main options for the protocol, and optionally an advanced button which opens one second dialog containing extra options.
- There is no more distinction between mandatory and optional arguments, but the
dialog ensures all mandatory arguments are handle, falling back to a generic 
dialog if they are not.

- It also simplifies the loading of configuration widgets by creating the 
AccountEditDialog, which does all the job of loading the dialogs. This dialog
is used by both the EditAccountDialog and the AddAccountWizard. It will also 
be used in the Kopete Telepathy plugin.

The port of the plugins to this new scheme is already being done.

svn path=/trunk/playground/network/telepathy-accounts-kcm/; revision=1181227