Compile against Telepathy-Qt4 0.5 Changed account abstract-account-parameters-widget.h to return QVariantMap of parameterValues() instead of trying to QMap<ParameterItem,QString> which didn't compile.
There didn't appear to be a reason for doing this as it was just copied
to a QVariantMap directly after anyway.

Note that the code now doesn't remove any parameters which have the same
value as the default. This didn't seem to be a step required by
Telepathy.

Also updated parameter-edit-model to always return the value in the type
requested by the ProtocolParameter.