Fix height for local-xmpp list view
The list item was slightly smaller than the containing list view for
small font sizes. When it was selected there was a small strip below
which didn't belong to the item and didn't change color. Fixed this.

For really large fonts the list view didn't get high enough and thus
got truncated. This gets fixed at the same time.

REVIEW: 108153
BUG: 309050