Avoid warning messages when closing the settings dialog
KWidgetItemDelegate doesn't like it if its view is deleted before
the delegate itself. This happens though if the delegate is
parented to the dialog itself, since the dialog will delete the UI
in its destructor (thus deleting the view), but the delegate is only
deleted by ~QObject's deleteChildren() call (which happens later).
Parenting the delegates to the dialog's main layout instead
prevents the problem.
This might fix crash bug 314400, but it's not clear if it will.
BUG:314400