[kaccounts] Disable the steam plugin
It turns out that the steam purple plugin is really hacky and actually
requires system symlinks to be done as it tries to dlopen a hardcoded
.so of gnome-keyring. This is baaad and until it's fixed to use proper
keyring APIs or using KWallet, this will be disabled for now.

It's still possible to install it if one wishes to, there's a _README
which has all the needed details.