[kaccounts] Don't crash on null profile
When the user selects an account for which an invalid Telepathy profile
exists, things should not crash but rather just warn the user.

As this plugin is now under string freeze, I've copied a string from the
old kcm (as it uses the same catalog), it's not ideal but this does not
violate string freeze and gives user at least a hint of what he should
do.

BUG: 344594